Move UseExpressionBodyForLambda to Analyzers layer
