+++ src/Features/Core/Portable/FeaturesResources.resx
@@ -1151,12 +1151,6 @@ This version used in: {2}</value>
   <data name="Add_member_name" xml:space="preserve">
     <value>Add member name</value>
   </data>
-  <data name="Use_block_body_for_lambda_expressions" xml:space="preserve">
-    <value>Use block body for lambda expressions</value>
-  </data>
-  <data name="Use_expression_body_for_lambda_expressions" xml:space="preserve">
-    <value>Use expression body for lambda expressions</value>
-  </data>
   <data name="Convert_to_linq_call_form" xml:space="preserve">
     <value>Convert to LINQ (call form)</value>
   </data>
