+++ src/Features/Core/Portable/xlf/FeaturesResources.ja.xlf
@@ -2895,16 +2895,6 @@ Zero-width positive lookbehind assertions are typically used at the beginning of
         <target state="translated">{0} の変性を更新するには、アプリケーションを再起動する必要があります。</target>
         <note />
       </trans-unit>
-      <trans-unit id="Use_block_body_for_lambda_expressions">
-        <source>Use block body for lambda expressions</source>
-        <target state="translated">ラムダ式にブロック本体を使用する</target>
-        <note />
-      </trans-unit>
-      <trans-unit id="Use_expression_body_for_lambda_expressions">
-        <source>Use expression body for lambda expressions</source>
-        <target state="translated">ラムダ式に式本体を使用する</target>
-        <note />
-      </trans-unit>
       <trans-unit id="Value_colon">
         <source>Value:</source>
         <target state="translated">値:</target>
