+++ src/Features/Core/Portable/xlf/FeaturesResources.zh-Hans.xlf
@@ -2895,16 +2895,6 @@ Zero-width positive lookbehind assertions are typically used at the beginning of
         <target state="translated">更新 {0} 的差异需要重新启动应用程序。</target>
         <note />
       </trans-unit>
-      <trans-unit id="Use_block_body_for_lambda_expressions">
-        <source>Use block body for lambda expressions</source>
-        <target state="translated">对 lambda 表达式使用块主体</target>
-        <note />
-      </trans-unit>
-      <trans-unit id="Use_expression_body_for_lambda_expressions">
-        <source>Use expression body for lambda expressions</source>
-        <target state="translated">对 lambda 表达式使用表达式正文</target>
-        <note />
-      </trans-unit>
       <trans-unit id="Value_colon">
         <source>Value:</source>
         <target state="translated">值:</target>
